% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nr_matches_per_probe.R
\name{get_nr_matches_per_probe}
\alias{get_nr_matches_per_probe}
\title{Get number of matches per probe}
\usage{
get_nr_matches_per_probe(matches)
}
\arguments{
\item{matches}{Matches (generated using the `map_probes` function).}
}
\description{
Get number of matches per probe
}
