% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OOB.R
\name{get_OOB}
\alias{get_OOB}
\title{Get out-of-band betas/intensities.}
\usage{
get_OOB(rgset, normalized = TRUE, keep = c("both", "beta", "intensity"))
}
\arguments{
\item{rgset}{RGset (minfi).}

\item{normalized}{Normalize intensity-values (using the wateRmelon 'nasen()' function)?}

\item{keep}{What to keep? ('both', 'intensity', 'beta')}
}
\description{
Get out-of-band betas/intensities.
}
