% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probe_overlaps.R
\name{get_probe_overlaps}
\alias{get_probe_overlaps}
\title{Calculate 3'-subsequence overlap between a set of probes}
\usage{
get_probe_overlaps(probes)
}
\arguments{
\item{probes}{Probes to test (450k/EPIC)}
}
\value{
A data frame with one row for each match and .. columns 
  \item{Probe_Index}{Probe ID}
  \item{Bead_Index}{Bead ID: Probe ID with '_methylated' or '_unmethylated' appended for type I probes}
  \item{Probe_Target}{Probe ID}
  \item{Bead_Target}{Bead ID: Probe ID with '_methylated' or '_unmethylated' appended for type I probes}
  \item{overlap}{overlap in bp}
  \item{mismatch_pos}{position of mismatch (bp from 3'end of probe), NA if exact match}
}
\description{
Calculate 3'-subsequence overlap between a set of probes
}
