% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{locusplot}
\alias{locusplot}
\title{Locusplot}
\usage{
locusplot(
  cpg,
  stats,
  windowsize,
  threshold = NULL,
  significance_line = FALSE,
  highlight = TRUE,
  text_size = 13,
  point_size = 2.5
)
}
\arguments{
\item{cpg}{CpG-site to plot}

\item{stats}{data.frame containing association p-values. Should contain a 'Probe' and a 'p' column.}

\item{windowsize}{Size (bp) of the window surrounding the CpG-site that should be plotted.}

\item{threshold}{Significance threshold. By default will use bonferonni threshold, using the number of probes in the `stats` file.}

\item{significance_line}{Plot a significane line(TRUE/FALSE)?}

\item{highlight}{Highlight significant sites?}

\item{text_size}{Text size.}

\item{point_size}{Point size.}
}
\description{
Locusplot
}
