% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probes.R
\name{map_probes}
\alias{map_probes}
\title{Map 450k/EPIC probes to the reference genome}
\usage{
map_probes(
  probes,
  path,
  array = "450k",
  chromosomes = "all",
  min_width = 15,
  max_width = 50,
  step_size = 5,
  allow_mismatch = FALSE,
  allow_INDEL = FALSE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{probes}{Probe names (450k or EPIC).}

\item{path}{directory were bisulfite-converted genomes are stored (see details).}

\item{array}{Array used (450k or EPIC)}

\item{chromosomes}{Which chromosomes to include (by default: chr1:22, X, Y and M).}

\item{min_width}{Minimum probe length to map (starting from the 3'-end of the probe).}

\item{max_width}{Maximum probe length to map (starting from the 3'-end of the probe).}

\item{step_size}{Map probe lengths from min_width to max_width in these steps.}

\item{allow_mismatch}{Allow a mismatch in matching? (TRUE/FALSE)}

\item{allow_INDEL}{Allow an INDEL in matching? (TRUE/FALSE)}

\item{cores}{Number of cores to use (default = 1).}

\item{verbose}{Verbose (TRUE/FALSE).}
}
\value{
A data frame with one row for each match and .. columns 
    \item{Probe}{Probe ID}
    \item{chr, start, end, strand}{chromosome positions}
    \item{next_base}{base preceding the 'C' base for type I probes}
    \item{mismatch_pos}{position of mismatch (bp from 3'end of probe), NA if exact match}
    \item{Type2}{Type: II, I_Methylated or I_Unmethylated}
    \item{width}{width (in basepairs) of the match}
    \item{channel}{predicted color channel for type I probes}
}
\description{
Map 450k/EPIC probes to the reference genome
}
\examples{
# Map some probes to all chromosomes, allowing no mismatch/INDELs
probes <- c("cg19883843", "cg24299847", "cg05542153")
\dontrun{
matches <- DNAmCrosshyb::map_probes(probes,
           path = "~/Desktop/genome_bs/hg19",
           chromosomes = "all",
           min_width = 30,
           max_width = 50,
           step_size = 5,
           allow_mismatch = FALSE,
           allow_INDEL = FALSE,
           cores = 1
            )
}
}
