% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probes_sequence.R
\name{map_probes_sequence}
\alias{map_probes_sequence}
\title{Map 450k/EPIC probes to a user-defined sequence}
\usage{
map_probes_sequence(
  sequence,
  next_base,
  prev_base,
  array = c("450k", "EPIC"),
  max_width = 50,
  min_width = 15,
  step_size = 5,
  allow_mismatch = FALSE,
  max_mismatch = 1,
  allow_indel = FALSE,
  min_distance = 6,
  use_Y = TRUE,
  methylation_status = "methylated",
  verbose = TRUE,
  cores = 1
)
}
\arguments{
\item{sequence}{DNA sequence (string)}

\item{next_base}{Base following the end of the sequence (necessary to know for bisulfite conversion)}

\item{prev_base}{Base preceding the start of the sequence (necessary to know for bisulfite conversion)}

\item{array}{Array used (450k or EPIC)}

\item{max_width}{Maximum probe length to map (starting from the 3'-end of the probe).}

\item{min_width}{Minimum probe length to map (starting from the 3'-end of the probe).}

\item{step_size}{Map probe lengths from min_width to max_width in these steps.}

\item{allow_mismatch}{Allow a mismatch in matching? (TRUE/FALSE)}

\item{max_mismatch}{Maximum number of allowed mismatches}

\item{allow_indel}{Allow an INDEL in matching? (TRUE/FALSE)}

\item{min_distance}{Minimum distance from 3'end of probe where mismatches/indels are allowed}

\item{use_Y}{Use Y (IUPAC) to represent Cs in CpG-sites?}

\item{methylation_status}{Assumed CpG-sites are either methylated or unmethylated (argument not used if use_Y == TRUE)}

\item{verbose}{Should function be verbose? (TRUE/FALSE)}

\item{cores}{Number of cores to use (default = 1).}
}
\value{
A data frame with one row for each match and .. columns 
  \item{Probe}{Probe ID}
  \item{start, end, strand}{positions}
  \item{width}{width (in basepairs) of the match}
  \item{sbe_site}{base preceding the match}
  \item{mismatch_pos}{position of mismatch (bp from 3'end of probe), NA if exact match}
  \item{indel_pos}{position of INDEL (bp from 3'end of probe), NA if exact match}
  \item{width_incl_indel}{width of match including INDEL, NA if exact match}
  \item{sequence_bs}{bisulfite-converted sequence}
  \item{Type2}{Type: II, I_Methylated or I_Unmethylated}
  \item{channel}{predicted color channel for type I probes}
}
\description{
Map 450k/EPIC probes to a user-defined sequence
}
\examples{
# Map probes to the C9orf72 hexanucleotide repeat
repeat_sequence <- paste(rep("GGCCCC", 10), collapse="")
matches_c9 <- map_probes_sequence(sequence = repeat_sequence, next_base = "G", prev_base = "C",
                                  array = "450k", min_width = 10, max_width = 25, allow_indel = FALSE, 
                                  allow_mismatch = TRUE, min_distance = 6,
                                  step_size = 1, use_Y = FALSE, methylation_status = "methylated")
head(matches_c9 \%>\% data.frame())
}
